refactor: class와 override를 이용하여 가위바위보와 묵찌빠 파일분리
